requirements.txt
streamlit==1.17.0
pytesseract==0.3.10
Pillow==9.2.0